+++ src/ui/components/sdcard.c
@@ -19,22 +19,39 @@
 #include <hardfault.h>
 #include <screen.h>
 #include <sd.h>
-#include <string.h>
 #include <touch/gestures.h>
 #include <ui/screen_stack.h>
 
 typedef struct {
-    // if true, the callback won't be called until the sd card is inserted.
-    // the insert/remove label changes depending on this flag.
-    bool insert;
     void (*callback)(bool, void*);
     void* callback_param;
+    // TODO: use a TIMER interrupt to get a more accurate timer.
+    // 250 is ~0.5 sec. Unit: rendering rate.
+    int check_interval;
+    int count;
 } data_t;
+
+static void _insert_poll_callback(component_t* component)
+{
+    data_t* data = (data_t*)component->data;
+    if (data->callback && sd_card_inserted()) {
+        data->callback(true, data->callback_param);
+        data->callback = NULL;
+        return;
+    }
+}
 
 static void _render(component_t* component)
 {
     image_sdcard(screen_is_upside_down());
     ui_util_component_render_subcomponents(component);
+
+    data_t* data = (data_t*)component->data;
+    if (data->count == data->check_interval) {
+        data->count = 0;
+        _insert_poll_callback(component);
+    }
+    data->count++;
 }
 
 /********************************** Component Functions **********************************/
@@ -47,17 +64,6 @@ static const component_functions_t _component_functions = {
 
 /********************************** Create Instance **********************************/
 
-static void _continue_callback(component_t* component)
-{
-    data_t* data = (data_t*)component->parent->data;
-    if (!data->insert || sd_card_inserted()) {
-        if (data->callback) {
-            data->callback(true, data->callback_param);
-            data->callback = NULL;
-        }
-    }
-}
-
 static void _cancel_callback(component_t* component)
 {
     data_t* data = (data_t*)component->parent->data;
@@ -67,7 +73,7 @@ static void _cancel_callback(component_t* component)
     }
 }
 
-component_t* sdcard_create(bool insert, void (*callback)(bool, void*), void* callback_param)
+component_t* sdcard_create(void (*callback)(bool, void*), void* callback_param)
 {
     component_t* component = malloc(sizeof(component_t));
     if (!component) {
@@ -80,9 +86,10 @@ component_t* sdcard_create(bool insert, void (*callback)(bool, void*), void* cal
     memset(data, 0, sizeof(data_t));
     memset(component, 0, sizeof(component_t));
 
-    data->insert = insert;
     data->callback = callback;
     data->callback_param = callback_param;
+    data->check_interval = 250;
+    data->count = 0;
     component->data = data;
     component->f = &_component_functions;
     component->dimension.width = SCREEN_WIDTH;
@@ -91,15 +98,11 @@ component_t* sdcard_create(bool insert, void (*callback)(bool, void*), void* cal
     ui_util_add_sub_component(
         component,
         label_create(
-            insert ? "Insert SD card\nto continue" : "Remove SD card\nto continue",
+            "Insert SD card\nto continue",
             NULL,
             screen_is_upside_down() ? RIGHT_CENTER : LEFT_CENTER,
             component));
     ui_util_add_sub_component(
-        component, icon_button_create(top_slider, ICON_BUTTON_CHECK, _continue_callback));
-    if (insert) {
-        ui_util_add_sub_component(
-            component, icon_button_create(top_slider, ICON_BUTTON_CROSS, _cancel_callback));
-    }
+        component, icon_button_create(top_slider, ICON_BUTTON_CROSS, _cancel_callback));
     return component;
 }