+++ src/ui/components/sdcard.h
@@ -22,6 +22,6 @@
  * @param[in] insert if true, the user is asked to insert the sdcard. Otherwise the user is asked to
  *            remove it.
  */
-component_t* sdcard_create(bool insert, void (*callback)(bool, void*), void* callback_param);
+component_t* sdcard_create(void (*callback)(bool, void*), void* callback_param);
 
 #endif